@@ -21,18 +21,19 @@
     "url": "https://github.com/digitalbazaar/bedrock-https-agent/issues"
   },
   "homepage": "https://github.com/digitalbazaar/bedrock-https-agent",
-  "dependencies": {},
+  "dependencies": {
+    "esm": "^3.2.25"
+  },
   "peerDependencies": {
-    "bedrock": "^4.3.0"
+    "bedrock": "^4.5.0"
   },
   "engines": {
-    "node": ">=8.9.0"
+    "node": ">=10"
   },
   "devDependencies": {
-    "eslint": "^7.3.1",
-    "eslint-config-digitalbazaar": "^2.0.0",
-    "eslint-plugin-jsdoc": "^22.0.0",
-    "jsdoc": "^3.5.5",
-    "jsdoc-to-markdown": "^5.0.3"
+    "eslint": "^7.13.0",
+    "eslint-config-digitalbazaar": "^2.8.0",
+    "eslint-plugin-jsdoc": "^37.9.7",
+    "jsdoc-to-markdown": "^6.0.1"
   }
 }
