@@ -1,9 +1,7 @@
 /*!
- * Copyright (c) 2019-2020 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-const {config} = require('bedrock');
+import {config} from 'bedrock';
 
 const cfg = config['https-agent'] = {};
 
